Obtain a further speedup by replacing CharBag vectors with GMP bignums.
The idea here is that we map the occurring Unicode character to the
smallest prime numbers, starting from the character that occurs the
most.

Now, the product of the mapped small primes represents the multiset.
Subset test is reduced to an integer divisibility test and multiset
subtraction is reduced to an integer division.