Reverse the order in which the words are considered.
A side effect of the way the charbags are hashed right now is that
small charbags tend to be hashed to small numbers, so the program
tends to recurse first on short words like "is". For long inputs, this
is a very tedious start.

It actually makes sense to do it in reverse. This also makes the
program output phrases much faster at the beginning when it slurps
very large charbags first (and correspondingly slower at the end).